Fix bug for unresponsive buttons
The millis were stored in a wrong data type, causing the
number to be negative after a certain period.

+++ EventListener.h
@@ -4,7 +4,7 @@
 class EventListener {
 private:
     int previousChoice;
-    int previousChoiceTime;
+    unsigned long previousChoiceTime;
     bool readButton(int value);
 public:
     EventListener(/* args */);
